getSomeYearsAgo hook created
