+++ client/src/pages/ServiceForm.jsx
@@ -5,6 +5,7 @@ import UserContext from '../contexts/UserContext';
 import ParkingDetailsContext from '../contexts/ParkingDetailsContext';
 import 'react-phone-number-input/style.css';
 import '../components/styles/Booking/ServiceForm.css';
+import getSomeYearsAgo from '../hooks/getSomeYearsAgo';
 
 function ServiceForm() {
   const { startDate, endDate } = useParams();
@@ -22,16 +23,6 @@ function ServiceForm() {
   const [dateOfBirth, setDateOfBirth] = useState('');
   const [phoneNumber, setPhoneNumber] = useState();
 
-  function getSomeYearsAgo(number) {
-    const date = new Date();
-    date.setFullYear(date.getFullYear() - number);
-
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-
-    return `${year}-${month}-${day}`;
-  }
 
   const handleBooking = async (event, userObj, parking) => {
     event.preventDefault();
